Fix minor proof bugs in the simple examples
